backend/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework >=3.10.3
flake8 >=3.7.9
Faker >= 3.0.0